cleanup and more tests
